Move posts functional tests to the separate directory
